+++ src/font.hpp
@@ -17,10 +17,10 @@ public:
     int ascent;
     char character;
     // advance depends on scale and must be declared after
-    float scale;
-    float advance;
+    double scale;
+    double advance;
 
-    FontCharacter(char c, float pixelHeight, char nextChar);
+    FontCharacter(char c, double pixelHeight, char nextChar);
     ~FontCharacter();
 
     FontCharacter(FontCharacter&&) = default;
@@ -33,23 +33,23 @@ public:
 class FontString {
 public:
     std::vector<FontCharacter> data;
-    float normalizedW;
-    float normalizedH;
+    double normalizedW;
+    double normalizedH;
 
-    auto static from_width(std::string_view string, float desiredPixelWidth) -> FontString;
-    auto static from_width_normalized(std::string_view string, float desiredWidth) -> FontString;
-    auto static from_height(std::string_view const string, float const desiredPixelHeight) -> FontString {
+    auto static from_width(std::string_view string, double desiredPixelWidth) -> FontString;
+    auto static from_width_normalized(std::string_view string, double desiredWidth) -> FontString;
+    auto static from_height(std::string_view const string, double const desiredPixelHeight) -> FontString {
         return FontString(string, desiredPixelHeight);
     }
-    auto static from_height_normalized(std::string_view const string, float const desiredHeight) -> FontString {
+    auto static from_height_normalized(std::string_view const string, double const desiredHeight) -> FontString {
         return from_height(string, get_window_dimensions().h * desiredHeight);
     }
-    auto static get_text_width(std::string_view text, float fontSize) -> float;
-    auto static get_text_width_normalized(std::string_view text, float fontSize) -> float;
+    auto static get_text_width(std::string_view text, double fontSize) -> double;
+    auto static get_text_width_normalized(std::string_view text, double fontSize) -> double;
 
 private:
-    FontString(std::string_view string, float pixelHeight);
+    FontString(std::string_view string, double pixelHeight);
 };
 
-auto get_codepoint_kern_advance(char codepoint, char nextCodepoint, float scale) -> float;
+auto get_codepoint_kern_advance(char codepoint, char nextCodepoint, double scale) -> double;
 auto init_font(std::string const& filePath) -> bool;
