+++ src/ui.hpp
@@ -12,35 +12,35 @@ namespace UI {
     // children's regions. These two classes make sure that values of one scale
     // aren't accidentally assigned to variables of the other scale.
     class WindowScale {
-        float value;
+        double value;
     public:
         WindowScale() noexcept = default;
-        constexpr explicit operator float() const { return value; };
-        constexpr WindowScale(float const from) noexcept : value{from} {};
+        constexpr explicit operator double() const { return value; };
+        constexpr WindowScale(double const from) noexcept : value{from} {};
 
         auto constexpr operator +=(WindowScale const& rhs) -> WindowScale& {
-            value += float(rhs);
+            value += double(rhs);
             return *this;
         }
         auto constexpr friend operator +(WindowScale lhs, WindowScale const& rhs) -> WindowScale {
             return lhs += rhs;
         }
         auto constexpr operator -=(WindowScale const& rhs) -> WindowScale& {
-            value -= float(rhs);
+            value -= double(rhs);
             return *this;
         }
         auto constexpr friend operator -(WindowScale lhs, WindowScale const& rhs) -> WindowScale {
             return lhs -= rhs;
         }
         auto constexpr operator *=(WindowScale const& rhs) -> WindowScale& {
-            value *= float(rhs);
+            value *= double(rhs);
             return *this;
         }
         auto constexpr friend operator *(WindowScale lhs, WindowScale const& rhs) -> WindowScale {
             return lhs *= rhs;
         }
         auto constexpr operator /=(WindowScale const& rhs) -> WindowScale& {
-            value /= float(rhs);
+            value /= double(rhs);
             return *this;
         }
         auto constexpr friend operator /(WindowScale lhs, WindowScale const& rhs) -> WindowScale {
@@ -49,35 +49,35 @@ namespace UI {
     };
 
     class RelativeScale {
-        float value;
+        double value;
     public:
         RelativeScale() noexcept = default;
-        constexpr explicit operator float() const { return value; };
-        constexpr RelativeScale(float const from) noexcept : value{from} {};
+        constexpr explicit operator double() const { return value; };
+        constexpr RelativeScale(double const from) noexcept : value{from} {};
 
         auto constexpr operator +=(RelativeScale const& rhs) -> RelativeScale& {
-            value += float(rhs);
+            value += double(rhs);
             return *this;
         }
         auto constexpr friend operator +(RelativeScale lhs, RelativeScale const& rhs) -> RelativeScale {
             return lhs += rhs;
         }
         auto constexpr operator -=(RelativeScale const& rhs) -> RelativeScale& {
-            value -= float(rhs);
+            value -= double(rhs);
             return *this;
         }
         auto constexpr friend operator -(RelativeScale lhs, RelativeScale const& rhs) -> RelativeScale {
             return lhs -= rhs;
         }
         auto constexpr operator *=(RelativeScale const& rhs) -> RelativeScale& {
-            value *= float(rhs);
+            value *= double(rhs);
             return *this;
         }
         auto constexpr friend operator *(RelativeScale lhs, RelativeScale const& rhs) -> RelativeScale {
             return lhs *= rhs;
         }
         auto constexpr operator /=(RelativeScale const& rhs) -> RelativeScale& {
-            value /= float(rhs);
+            value /= double(rhs);
             return *this;
         }
         auto constexpr friend operator /(RelativeScale lhs, RelativeScale const& rhs) -> RelativeScale {
